feat(trainer): start service fastapi and worker
