+++ trainer_service/app/api/routers/__init__.py
@@ -1,8 +1,7 @@
 from fastapi import APIRouter
-from .trainer import router
 
 api_routers = APIRouter()
 
-api_routers.include_router(router)
-
-__all__ = ['api_routers']
+@api_routers.get("/health")
+async def health_check():
+    return {"status": "ok"}
