+++ trainer_service/fastapi_app.py
@@ -1,0 +1,6 @@
+from fastapi import FastAPI
+from app.api.routers import api_routers
+
+app = FastAPI()
+
+app.include_router(api_routers)
