+++ trainer_service/main.py
@@ -1,56 +1,15 @@
-import os
-import httpx
+import uvicorn
 import asyncio
-from app.core import training_model
-from app.logs import get_logger
+from fastapi_app import app
+from worker import worker_loop
 
-TASKER_URL = "http://" + os.getenv("TASKER", "localhost:6110")
-
-logger = get_logger(__name__)
-
-async def worker_loop():
-    async with httpx.AsyncClient(timeout=30.0) as client:
-        logger.info("Сервис тренировки начал работу.")  
-        while True:
-            # 1. Запрашиваем следующую задачу
-            try:
-                resp = await client.post(f"{TASKER_URL}/tasks/next")
-            except Exception as e:
-                logger.error("Ошибка подключения к сервису задач")
-                continue
-            
-            if resp.status_code == 204 or resp.status_code == 404 or resp.text == "null" or not resp.content:
-                logger.info("Задач нет. Отдыхаем.")
-                await asyncio.sleep(1)
-                continue
-
-            task = resp.json()
-            if not task:
-                await asyncio.sleep(1)
-                continue
-            
-            task_id = task["task_id"]
-            payload = task["payload"]
-            print(f"Worker: взял задачу {task_id}")
-            
-            # 2. Обновляем статус: запущено
-            await client.put(f"{TASKER_URL}/tasks/{task_id}/status", json={"status": "running", "progress": 0})
-            
-            try:
-                
-                # Процесс обучения
-                training_model(task_id, payload["params_train"])
-
-                # Завершение
-                result = {"processed": payload, "message": "success"}
-                await client.put(f"{TASKER_URL}/tasks/{task_id}/status", 
-                                    json={"status": "completed", "progress": 100, "result": result})
-                print(f"Worker: задача {task_id} завершена")
-                
-            except Exception as e:
-                await client.put(f"{TASKER_URL}/tasks/{task_id}/status",
-                                 json={"status": "failed", "progress": 0, "result": {"error": str(e)}})
-                print(f"Worker: ошибка {task_id}: {e}")
+async def main():
+    server = uvicorn.Server(uvicorn.Config(app, host="0.0.0.0", port=6200))
+    
+    await asyncio.gather(
+        server.serve(),
+        worker_loop()
+    )
 
 if __name__ == "__main__":
-    asyncio.run(worker_loop())
+    asyncio.run(main())
